refactor(kanban): extract notify-subs + worker-spawn mixins from kanban_db.py (shard s5)
